backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2

langchain==0.2.7
langchain-community==0.2.6

chromadb==0.5.3
unstructured==0.14.9
pdfminer.six==20231228
tiktoken==0.7.0
